attemps and converging astro and vite parts
