+++ .vscode/settings.json
@@ -1,4 +1,4 @@
 {
   "editor.formatOnSave": true,
-  "cSpell.words": ["Astro", "Astro's", "spotlightjs"]
+  "cSpell.words": ["Astro", "Astro's", "contextlines", "spotlightjs"]
 }
