+++ demos/sveltekit/src/hooks.server.ts
@@ -1,7 +1,6 @@
 import * as Sentry from '@sentry/sveltekit';
 
 Sentry.init({
-	debug: true,
 	tracesSampleRate: 1.0,
 	spotlight: import.meta.env.DEV
 });
