+++ packages/astro/src/snippets.ts
@@ -37,27 +37,17 @@ ${buildClientInit(options)}
  * Hook into Vite's client code to enable Spotlight if an error occurs.
  *
  * Main difference to normal page injection: We only initialize spotlight
- * if an error happens and is transmitted via the websocket. This is to avoid
- * initializing this spotlight instance before the actual app's instance.
- * This should only be the fallback instance if the app failed to initialize.
+ * if an error happens.
  *
- * Used variables from Vite's client code:
- * - `enableOverlay` to check if the overlay should be shown
- *    @see https://github.com/vitejs/vite/blob/b9ee620108819e06023e4303af75a61d3e4e4d76/packages/vite/src/client/client.ts#L289
- * - `socket` to listen for incoming error events:
- *    @see https://github.com/vitejs/vite/blob/b9ee620108819e06023e4303af75a61d3e4e4d76/packages/vite/src/client/client.ts#L67
+ * Hacked `createErrorOverlay` from Vite's client code:
+ * @see https://github.com/vitejs/vite/blob/b9ee620108819e06023e4303af75a61d3e4e4d76/packages/vite/src/client/client.ts#L291
  */
 export const buildSpotlightErrorPageSnippet = (options: ClientInitOptions) => `
 ${buildClientImport(options.importPath)}
 
-if (enableOverlay) {
-  socket.addEventListener('message', (event) => {
-    const dataJson = JSON.parse(event.data);
-    if (dataJson.type === 'error') {
-      ${buildClientInit(options)}
-    }
-  });
-}
+createErrorOverlay = (err) => {
+  ${buildClientInit(options)}
+};
 `;
 
 export const buildServerSnippet: (options: SpotlightAstroIntegrationOptions) => string = options => `
